:truck: chore(api) move audit logger repository to `src/`
