+++ api/src/identity-access-management/application/jobs/event-logging.job-controller.js
@@ -1,6 +1,6 @@
-import { auditLoggerRepository } from '../../../../lib/infrastructure/repositories/audit-logger-repository.js';
 import { JobController } from '../../../shared/application/jobs/job-controller.js';
 import { EventLoggingJob } from '../../domain/models/jobs/EventLoggingJob.js';
+import { auditLoggerRepository } from '../../infrastructure/repositories/audit-logger-repository.js';
 
 export class EventLoggingJobController extends JobController {
   constructor() {
