+++ api/tests/identity-access-management/integration/infrastructure/repositories/audit-logger-repository_test.js
@@ -1,8 +1,8 @@
-import { httpAgent } from '../../../../lib/infrastructure/http/http-agent.js';
-import { auditLoggerRepository } from '../../../../lib/infrastructure/repositories/audit-logger-repository.js';
-import { config } from '../../../../src/shared/config.js';
-import { AuditLoggerApiError } from '../../../../src/shared/domain/errors.js';
-import { catchErr, expect, sinon } from '../../../test-helper.js';
+import { httpAgent } from '../../../../../lib/infrastructure/http/http-agent.js';
+import { auditLoggerRepository } from '../../../../../src/identity-access-management/infrastructure/repositories/audit-logger-repository.js';
+import { config } from '../../../../../src/shared/config.js';
+import { AuditLoggerApiError } from '../../../../../src/shared/domain/errors.js';
+import { catchErr, expect, sinon } from '../../../../test-helper.js';
 
 const { auditLogger } = config;
